Add proper PBR/HDR post-processing pass
First, an exposure coefficient is applied, then an optional Purkinje
effect, then a tone-mapping function to go
from HDR to SDR (accounting for desired dynamic range).

Exposure can be automatically set based on the scene's average
(weighted) luminance.

Optionally, a bloom effect can be performed, on a high-pass filter of
the resulting scene.

At this stage, any color value which has a component greater than 1 is rescaled down to
fit the [0; 1] range.